requirements.txt
django>=4.0.0
djangorestframework
djongo==1.3.6
pymongo==3.12.3
django-cors-headers
requests
pyyaml
